created two sets of buttons, depending if the user is an admin or not and displaying it accordingly
